Cardio good fitness
Context
Explore this dataset to identify differences between the customers of each product. Also to explore relationships between the different attributes of the customers.
Data Description:¶
The data is about customers of the treadmill product(s) of a retail store called Cardio Good Fitness. It contains the following variables:

Product - The model no. of the treadmill

Age - Age of the customer in no of years

Gender - Gender of the customer

Education - Education of the customer in no. of years

Marital Status- Marital status of the customer

Usage - Avg. # times the customer wants to use the treadmill every week

Fitness - Self rated fitness score of the customer (5 - very fit, 1 - very unfit)

Income - Income of the customer

Miles- Miles that a customer expects to run